improve vps resource structure so the request flow is in a separated file
